+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v3.0.0] - 2019-11-03
+
 ### Added
 
 *   Status bar control.
@@ -44,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 *   Stage controller based on view controller.
 
 
-[Unreleased]: https://github.com/falkoschumann/java-muspellheim-commons-fx/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/falkoschumann/java-muspellheim-commons-fx/compare/v3.0.0...HEAD
+[v3.0.0]: https://github.com/falkoschumann/java-muspellheim-commons-fx/compare/v2.0.0...v3.0.0
 [v2.0.0]: https://github.com/falkoschumann/java-muspellheim-commons-fx/compare/v1.0.0...v2.0.0
 [v1.0.0]: https://github.com/falkoschumann/java-muspellheim-commons-fx/tree/v1.0.0